Minor adjustments to mock question
