+++ src/main/java/spring/certification/test/q005/example/Blood.java
@@ -1,0 +1,16 @@
+package spring.certification.test.q005.example;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.stereotype.Component;
+
+/**
+ * Red liquid.
+ */
+@Setter
+@Getter
+@Component
+public class Blood {
+    private int oxygenRatio;
+    private int carbonDioxideRatio;
+}
